+++ docs/changelog.md
@@ -1,4 +1,7 @@
 # TBD
+### Fixes
+* Made the API container host machine info optional, to represent that host machine info will only be present if the API container is running
+* `EnclaveContext`s cannot be constructed if the API container isn't running
 
 # 0.7.1
 ### Fixes